# protobuf/gen/openapi/services/product_resource/v1/openapi.yaml
# File generated by protoc-gen-pocket-openapi. DO NOT EDIT.
openapi: 3.0.0
info:
  title: product-resource
  version: "1.0.0"



paths:
  /product-resource/v1/products:
    post:
      summary: Cria um produto
      description: Realiza a criação de um novo produto.
      operationId: CreateProduct
      requestBody:
        required: true
        description: 
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        '201':
          description: Sucesso.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateProductResponse"
        '400':
          description: Argumentos da requisição inválidos.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
        '500':
          description: Erro interno do servidor.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DefaultError"
  /product-resource/v1/products/{id}:
    delete:
      summary: Remove um produto
      description: Faz a remoção de um produto, certificando de não remover um produto já removido previamente.
      operationId: DeleteProduct
      parameters:
        - in: path
          name: "id"
          required: true
          description: ID do produto
          schema:
            type: string
            example: prd_42v0t5IUiw9iIxpddWnybMCleq7uMynCbANK1VTxaT
      responses:
        '200':
          description: Sucesso.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteProductResponse"
        '400':
          description: Argumentos da requisição inválidos.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
        '404':
          description: Documento não encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DefaultError"
        '500':
          description: Erro interno do servidor.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DefaultError"
    get:
      summary: Procura por um produto
      description: Realiza a procura por um produto através de seu ID. Não exibindo produtos removidos.
      operationId: GetProductByID
      parameters:
        - in: path
          name: "id"
          required: true
          description: ID do produto
          schema:
            type: string
            example: prd_42v0t5IUiw9iIxpddWnybMCleq7uMynCbANK1VTxaT
        - in: query
          name: "return_deleted"
          required: false
          description: Quando true indica que produtos removidos serão retornados
          schema:
            type: boolean
      responses:
        '200':
          description: Sucesso.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetProductByIDResponse"
        '400':
          description: Argumentos da requisição inválidos.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
        '404':
          description: Documento não encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DefaultError"
        '500':
          description: Erro interno do servidor.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DefaultError"
    put:
      summary: Atualiza um produto
      description: Realiza a atualização de um produto através de seu ID.
      operationId: UpdateProduct
      parameters:
        - in: path
          name: "id"
          required: true
          description: ID do produto
          schema:
            type: string
            example: prd_42v0t5IUiw9iIxpddWnybMCleq7uMynCbANK1VTxaT
      requestBody:
        required: false
        description: 
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductRequest"
      responses:
        '200':
          description: Sucesso.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateProductResponse"
        '400':
          description: Argumentos da requisição inválidos.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
        '404':
          description: Documento não encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DefaultError"
        '500':
          description: Erro interno do servidor.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DefaultError"

components:
  schemas:
    CreateProductRequest:
      type: object
      properties:
        name:
          type: string
          description: Nome do produto que será criado
          example: Nome
    CreateProductResponse:
      type: object
      properties:
        product:
          $ref: '#/components/schemas/ProductProto'
    DefaultError:
      type: object
      properties:
        errors:
          type: array
          items:
            type: string
        message:
          type: string
    DeleteProductResponse:
      type: object
      properties:
        product:
          $ref: '#/components/schemas/ProductProto'
    FieldValidationError:
      type: object
      properties:
        field:
          type: string
        location:
          type: string
        message:
          type: string
    GetProductByIDResponse:
      type: object
      properties:
        product:
          $ref: '#/components/schemas/ProductProto'
    ProductProto:
      type: object
      properties:
        created_at:
          type: string
          format: date-time
          description: Data/hora de criação do produto
          example: "2023-12-14T16:28:44Z"
        deleted_at:
          type: string
          format: date-time
          description: Data/hora da remoção do produto
          example: "2023-12-14T16:28:44Z"
        id:
          type: string
          description: ID do produto
          example: prd_42v0t5IUiw9iIxpddWnybMCleq7uMynCbANK1VTxaT
        name:
          type: string
          description: Nome do produto
          example: Action figure
        updated_at:
          type: string
          format: date-time
          description: Data/hora da atualização do produto
          example: "2023-12-14T16:28:44Z"
    UpdateProductRequest:
      type: object
      properties:
        name:
          type: string
          description: Novo nome do produto
          example: Novo nome
    UpdateProductResponse:
      type: object
      properties:
        product:
          $ref: '#/components/schemas/ProductProto'
    ValidationError:
      type: object
      properties:
        errors:
          type: array
          items:
            $ref: '#/components/schemas/FieldValidationError'
        message:
          type: string 
  securitySchemes:
    authorization:
      type: http
      scheme: unspecified